refactor: 입출력 양식에 맞게 개행 문자 추가
입출력 양식에 맞추고자 개행 문자 및 출력문을 추가함

변경된 점:
- BonusNumberInput, LottoPurchaseAmountInput, WinningNumbersInput에 sout 구문을 추가
