+++ src/main/java/lotto/input/BonusNumberInput.java
@@ -15,6 +15,7 @@ public class BonusNumberInput {
         System.out.println("보너스 번호를 입력해 주세요.");
         String bonusNumberInput = Console.readLine();
         bonusNumberValidator.validate(bonusNumberInput);
+        System.out.println(); // 입출력을 맞추기 위한 개행
         return Integer.parseInt(bonusNumberInput);
     }
 }
