+++ src/main/java/lotto/input/WinningNumbersInput.java
@@ -1,7 +1,7 @@
 package lotto.input;
 
 import camp.nextstep.edu.missionutils.Console;
-import lotto.Lotto;
+import lotto.model.Lotto;
 import lotto.input.validator.WinningNumbersValidator;
 import lotto.util.ListIntegerConverter;
 
@@ -19,6 +19,7 @@ public class WinningNumbersInput {
         System.out.println("당첨 번호를 입력해 주세요.");
         String winningNumbersInput = Console.readLine();
         winningNumbersValidator.validate(winningNumbersInput);
+        System.out.println(); // 입출력을 맞추기 위한 개행
 
         List<Integer> winningNumbers = ListIntegerConverter.convertFromString(winningNumbersInput);
         return new Lotto(winningNumbers);
